apps/requirements.txt
alabaster==0.7.6
anaconda-client==1.2.1
argcomplete==1.0.0
astropy==1.0.6
Babel==2.3.4
beautifulsoup4==4.4.1
bitarray==0.8.1
blaze==0.8.3
bokeh==0.10.0
boto==2.38.0
Bottleneck==1.0.0
cffi==1.2.1
clyent==1.2.0
colorama==0.3.3
comtypes==1.1.2
conda==4.1.6
conda-build==1.18.2
conda-env==2.5.0a0
configobj==5.0.6
coverage==4.1
cryptography==1.0.1
cycler==0.9.0
Cython==0.23.4
cytoolz==0.7.4
datashape==0.4.7
decorator==4.0.6
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.10
django-debug-toolbar==1.4
django-phonenumber-field==1.1.0
django-simple-history==1.8.1
django-toolbelt==0.0.1
django-webpack-loader==0.3.0
djangorestframework==3.3.2
docutils==0.12
dynd===5a2f274
fastcache==1.0.2
Flask==0.10.1
greenlet==0.4.9
gunicorn==19.6.0
h5py==2.5.0
httpie==0.9.4
idna==2.0
ipykernel==4.2.2
ipython==4.0.3
ipython-genutils==0.1.0
ipywidgets==4.1.0
itsdangerous==0.24
jdcal==1.0
jedi==0.9.0
Jinja2==2.8
jsonschema==2.4.0
jupyter==1.0.0
jupyter-client==4.1.1
jupyter-console==4.0.3
jupyter-core==4.0.6
llvmlite==0.8.0
lxml==3.6.0
markdown2==2.3.1
MarkupSafe==0.23
matplotlib==1.5.1
menuinst==1.3.2
mistune==0.7.1
multipledispatch==0.4.8
nbconvert==4.1.0
nbformat==4.0.1
networkx==1.10
nltk==3.1
nose==1.3.7
notebook==4.1.0
numba==0.22.1
numexpr==2.4.4
numpy==1.10.1
odo==0.3.4
openpyxl==2.2.6
pandas==0.17.1
path.py==0.0.0
patsy==0.4.0
pep8==1.6.2
phonenumberslite==7.4.5
pickleshare==0.5
Pillow==3.0.0
ply==3.8
psutil==3.3.0
psycopg2==2.6.1
py==1.4.30
pyasn1==0.1.9
pycosat==0.6.1
pycparser==2.14
pycrypto==2.6.1
pycurl==7.19.5.1
pyflakes==1.0.0
Pygments==2.0.2
pykml==0.1.0
pyOpenSSL==0.15.1
pyparsing==2.0.3
pyreadline==2.1
pytest==2.8.1
python-dateutil==2.4.2
pytz==2016.6.1
pywin32==219
PyYAML==3.11
pyzmq==15.2.0
qtconsole==4.1.1
requests==2.9.1
rope-py3k==0.9.4.post1
ruamel-yaml===-VERSION
scikit-image==0.11.3
scikit-learn==0.17
scipy==0.16.0
seaborn==0.7.0
simplegeneric==0.8.1
six==1.10.0
snowballstemmer==1.2.0
sockjs-tornado==1.0.1
Sphinx==1.3.1
sphinx-rtd-theme==0.1.7
spyder==2.3.8
SQLAlchemy==1.0.9
sqlparse==0.2.0
static3==0.7.0
statsmodels==0.6.1
sympy==0.7.6.1
tables==3.2.2
toolz==0.7.4
tornado==4.3
traitlets==4.1.0
ujson==1.33
unicodecsv==0.14.1
virtualenv==13.0.1
Werkzeug==0.11.2
whitenoise==3.2
xlrd==0.9.4
XlsxWriter==0.8.4
xlwings==0.5.0
xlwt==1.0.0
